// src/enums/QueryKey.ts
export enum QueryKey {
  GET_USER = 'get-user',
  GET_CLIENTS = 'get-clients',
  GET_CLIENT_BY_ID = 'get-client-by-id',
  GET_LOCATIONS = 'get-locations',
  GET_AREA_BY_ID = 'get-area-by-id',
  GET_CONSTRUCTION_BY_ID = 'get-construction-by-id',
  GET_EQUIPMENT_BY_ID = 'get-equipment-by-id',
  GET_CHECKLISTS = 'get-checklists',
  GET_CHECKLIST_BY_ID = 'get-checklist-by-id',
  GET_USERS = 'get-users',
  GET_USER_BY_ID = 'get-user-by-id',
}